Fix styling bug with player ratings hidden
If player ratings are hidden (in lichess preferences) then lichess will
use grid area to expand the winners section to cover the empty space
left by the leaderboard.

Since in our styling the two sections are on either side of the daily
puzzle - this is not possible without UI glitches.